tests & refactor
